+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Improve the chance of surviving discord updates
+* Open reactions page on +(number) chip click